+++ tests/test_bot.py
@@ -126,6 +126,42 @@ class TestBot:
         assert "recommendation is available" in body
         print(body)
 
+    def test_reply_with_cc_and_to_bot(self, bcmd, ac_sender):
+        send_adr = ac_sender.adr
+        msg = mime.gen_mail_msg(
+            From=send_adr, To=[bcmd.bot_adr],
+            Cc=['some@address.example'],
+            Autocrypt=ac_sender.ac_headerval,
+            Subject="hello", _dto=True)
+
+        out = bcmd.run_ok(["bot-reply"], input=msg.as_string())
+        reply_msg = mime.parse_message_from_string(out)
+        assert reply_msg["To"] == send_adr
+        assert reply_msg["Cc"] == msg["Cc"]
+
+    def test_reply_with_cc_bot(self, bcmd, ac_sender):
+        send_adr = ac_sender.adr
+        msg = mime.gen_mail_msg(
+            From=send_adr, To=["some@address.example"],
+            Cc=[bcmd.bot_adr],
+            Autocrypt=ac_sender.ac_headerval,
+            Subject="hello", _dto=bcmd.bot_adr)
+
+        out = bcmd.run_ok(["bot-reply"], input=msg.as_string())
+        assert not out
+
+    def test_reply_puts_to_addresses_to_cc(self, bcmd, ac_sender, linematch):
+        send_adr = ac_sender.adr
+        msg = mime.gen_mail_msg(
+            From=send_adr, To=["bot2@example.org", bcmd.bot_adr],
+            Autocrypt=ac_sender.ac_headerval,
+            Subject="hello", _dto=bcmd.bot_adr)
+
+        out = bcmd.run_ok(["bot-reply"], input=msg.as_string())
+        reply_msg = mime.parse_message_from_string(out)
+        assert reply_msg["To"] == send_adr
+        assert reply_msg["Cc"] == "bot2@example.org"
+
     def test_reply_to_encrypted(self, bcmd, ac_sender, linematch):
         send_adr = ac_sender.adr
         msg = mime.gen_mail_msg(